.clang-tidy: Create lint config and apply
Apply for a random selection of files and warnings, making some
style fixes. Should probably metholodically go through each source
file and fix or ignore all releveant warnings. Will be tuning this
lint step as I go along.

+++ src/fwd.hpp
@@ -48,7 +48,7 @@ struct AxonOutModel;
 struct Message;
 
 enum BufferPosition : int;
-enum NeuronStatus : int;
+enum NeuronStatus : uint8_t;
 
 }
